+++ ad-search.php
@@ -305,7 +305,7 @@ if(isset($_POST['mostrarTodosAnuncios'])){
             </div>
         </section>
     </main>
-    <?php include "./footer.html" ?>
+    <?php include "./footer.php" ?>
 </body>
 
 </html>