+++ index.php
@@ -78,7 +78,7 @@ session_start();
     </section>
   </main>
 
-  <?php include "./footer.html" ?>
+  <?php include "./footer.php" ?>
 
   <script src="assets/bootstrap/js/bootstrap.min.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/baguettebox.js/1.10.0/baguetteBox.min.js"></script>