+++ legal-warning.php
@@ -57,7 +57,7 @@
         </div>
     </section>
 </main>
-<?php include "./footer.html" ?>
+<?php include "./footer.php" ?>
 <script src="assets/js/jquery.min.js"></script>
 <script src="assets/bootstrap/js/bootstrap.min.js"></script>
 <script src="https://cdnjs.cloudflare.com/ajax/libs/baguettebox.js/1.10.0/baguetteBox.min.js"></script>