+++ registration.php
@@ -162,7 +162,7 @@ if (isset($_POST['registrarUsuario'])) {
             </div>
         </section>
     </main>
-    <?php include "./footer.html" ?>
+    <?php include "./footer.php" ?>
 </body>
 
 </html>